+++ NEWS.md
@@ -9,6 +9,10 @@
 
 * new parameter `scale` added to `ph_with.gg`, `body_add_gg` and `body_add.gg` 
 to set the scale of ggplot outputs (like in ggsave).
+* new function `set_autonum_bookmark()` to recycle an object 
+made by `run_autonum()` by changing the bookmark value.
+* add `title_num_depth` argument to prefix a autonumber with the 
+title number (i.e. 4.3-2 for figure 2 of chapter 4.3).
 
 ## Deprecation